pyproject.toml
[project]
name = "yeti-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.9.0",
    "yeti-python>=2.0.8",
]

[tool.uv.sources]
yeti-python = { git = "https://github.com/yeti-platform/yeti-python.git", branch = "main" }

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
